Some files were relocated from folders for Redux components.
