+++ package.json
@@ -7,6 +7,7 @@
     "react-dom": "^15.6.1",
     "react-fa": "^4.2.0",
     "react-fontawesome": "^1.6.1",
+    "react-redux": "^5.0.6",
     "react-router-dom": "^4.1.2",
     "react-scripts": "1.0.11"
   },
