+++ src/js/pages/common/components/pageHeader.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import PageTitle from '../components/pageTitle';
-import SearchBox from '../components/searchBox';
+import PageTitle from './pageTitle';
+import SearchBox from './searchBox';
 import BorderedButton from '../../../common/components/controls/borderedButton';
 import '../../../../styles/page/__header/page__header.css'
