+++ src/js/pages/newSurveyPage/components/newSurveyPage.js
